feat(api): split server definition
